Fix bugs
FFmpeg, RTSP

+++ NabiSquareHD/ViewControllers/MediaManager/MediaManagerCollectionViewCell.m
@@ -87,7 +87,7 @@
     [operation setOutputStream:[NSOutputStream outputStreamToFileAtPath:fullPath append:NO]];
     
     [operation setDownloadProgressBlock:^(NSUInteger bytesRead, long long totalBytesRead, long long totalBytesExpectedToRead) {
-        NSLog(@"bytesRead: %u, totalBytesRead: %lld, totalBytesExpectedToRead: %lld", bytesRead, totalBytesRead, totalBytesExpectedToRead);
+        NSLog(@"bytesRead: %lu, totalBytesRead: %lld, totalBytesExpectedToRead: %lld", (long)bytesRead, totalBytesRead, totalBytesExpectedToRead);
         progressView.progressCounter = (bytesRead / totalBytesRead);
     }];
